Combined the award processing into a single class, and cleaned up some of the usage.
